Moved easyopencv samples to FtcRobotController, made the ftclib vision package have a skystone detector pipeline and a opmode that uses it
